+++ src/pages/auth/LoginPage.tsx
@@ -1,13 +1,6 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { toast } from 'react-toastify';
-import Input from '@/components/input/Input';
 import Button from '@/components/button/Button';
-import { LoginFormInputs } from '@/interfaces/auth.interfaces';
 import {
   Contents,
-  ErrorMessage,
   Form,
   InputContainer,
   LinkContainer,
@@ -15,75 +8,46 @@ import {
   Span,
   Title,
 } from '@/styles/AuthForm.style';
-import { getMyInfo, login } from '@/apis/auth.api';
-import { useSetAtom } from 'jotai';
-import { tokenAtom, userAtom } from '@/atoms/auth.atom';
+
+import { useLogin } from '@/hook/useLogin';
+import FormInput from '@/components/input/FormInput';
 
 export default function LoginPage() {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<LoginFormInputs>();
-  const navigate = useNavigate();
-
-  const setToken = useSetAtom(tokenAtom);
-  const setUser = useSetAtom(userAtom);
-
-  // 로딩상태
-  const [isLoading, setIsLoading] = useState(false);
-
-  const onSubmit: SubmitHandler<LoginFormInputs> = async data => {
-    setIsLoading(true);
-    try {
-      const { accessToken } = await login(data);
-      setToken(accessToken);
-
-      const userInfo = await getMyInfo();
-      setUser(userInfo);
-
-      toast.success(`${userInfo.nickname}님, 환영합니다!`);
-      navigate('/');
-    } catch (error) {
-      setToken(null);
-      setUser(null);
-
-      console.error('로그인 처리중 에러 발생:', error);
-      toast.error('로그인 실패');
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const { register, handleSubmit, errors, isValid, isLoading } = useLogin();
 
   return (
     <Contents>
       <Title>로그인</Title>
-      <Form onSubmit={handleSubmit(onSubmit)}>
+      <Form onSubmit={handleSubmit}>
         <InputContainer>
-          <Input
-            hasError={!!errors.email}
+          <FormInput
+            label="이메일"
+            name="email"
             type="email"
             placeholder="이메일을 입력하세요"
-            {...register('email', {
+            register={register}
+            errors={errors}
+            rules={{
               required: '이메일을 입력하세요.',
-            })}
+            }}
           />
-          {errors.email && <ErrorMessage>{errors.email.message}</ErrorMessage>}
 
-          <Input
-            hasError={!!errors.password}
+          <FormInput
+            label="비밀번호"
+            name="password"
             type="password"
             placeholder="비밀번호를 입력하세요"
-            {...register('password', {
+            register={register}
+            errors={errors}
+            rules={{
               required: '비밀번호를 입력하세요.',
-            })}
+            }}
           />
-          {errors.password && <ErrorMessage>{errors.password.message}</ErrorMessage>}
 
-          {/* 로딩 상태에 따라 버튼 비활성화 */}
-          <Button type="submit" size="large" disabled={isLoading}>
-            로그인
+          <Button type="submit" size="large" disabled={!isValid || isLoading}>
+            {isLoading ? '로그인 중...' : '로그인'}
           </Button>
+
           <LinkContainer>
             <Span>아직 회원이 아니신가요?</Span>
             <LinkStyle to={'/signup'}>회원가입</LinkStyle>
